backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
transformers==4.35.0
datasets==2.14.0
numpy==1.24.3
python-docx==1.1.0
PyPDF2==3.0.1
